requirements.txt
numpy
matplotlib
andi-datasets
pandas
ipykernel
torch 
torchvision 
torchaudio
mlflow
keras-tuner
pydot
tensorrt
deeptrack
cupy-cuda12x
